refactor(eat): removed duplicate code
