+++ modules/fbs-eat/eat/api/dashboard/dashboard.py
@@ -23,6 +23,17 @@ from api.util.utilities import (
     filter_data,
     convert_time,
     create_invisible_time_row,
+    filter_attempt_limits,
+    select_all_exercises_in_course,
+    filter_time,
+)
+from api.util.dashboard_util import (
+    hide_histogram,
+    create_course_bars,
+    create_figure_with_subfigure,
+    reduce_data_to_necessary_columns,
+    create_average_bar,
+    get_values_from_data,
 )
 
 tmp_df = get_data(-1)
@@ -59,6 +70,7 @@ layout = html.Div(
                                             ["Typical Mistakes", "Average Attempts"],
                                             "Typical Mistakes",
                                             style={"background-color": "#e9e7e9"},
+                                            clearable=False,
                                         ),
                                     ]
                                 ),
@@ -312,10 +324,6 @@ def update_histogram_avg_submissions(
     local_df["Time"] = pd.to_datetime(local_df["Time"])
     local_df = local_df[local_df["UserId"] != 0]
 
-    date_time_from = dateutil.parser.parse(date_time_from)
-
-    date_time_to = dateutil.parser.parse(date_time_to)
-
     if "Average Attempts" in key_figure_value:
         display_style = {"display": "block"}
     else:
@@ -323,9 +331,7 @@ def update_histogram_avg_submissions(
 
     task_len = {}
     if "Date" in checklist_value:
-        local_df = local_df[
-            (local_df.Time >= date_time_from) & (local_df.Time < date_time_to)
-        ]
+        local_df = filter_time(local_df, convert_time(date_time_from), convert_time(date_time_to))
     if not course_value or local_df.empty:
         return generate_empty_response(), display_style
 
@@ -351,6 +357,8 @@ def update_histogram_avg_submissions(
 
     return fig, display_style
 
+
+# pylint: disable=too-many-locals
 
 # Update histogramm figure
 @callback(
@@ -387,58 +395,37 @@ def update_histogram(
     :param daten: all data
     :return: updated histogram
     """
-
-    if "Typical Mistakes" in key_figure_value:
-        display_style = {"display": "block"}
-    else:
-        display_style = {"display": "none"}
+    # pylint: disable=duplicate-code
+    labels = [
+        "Joins",
+        "Projection_Attributes",
+        "Selection_Attributes",
+        "GroupBy",
+        "OrderBy",
+        "Strings",
+        "Tables",
+    ]
+    # pylint: enable=duplicate-code
+    display_style = hide_histogram(key_figure_value)
 
     local_df = filter_data(daten)
-    # Convert datetime string to datetime object
-    date_time_from, date_time_to = convert_time(date_time_from, date_time_to)
-
-    slider_value = slider_value or [local_df.Attempt.min(), local_df.Attempt.max()]
 
     # if no excercises are choosen, use all excercises
     if not exercise_value:
-        if course_value:
-            exercise_value = list(
-                local_df[local_df.CourseName.isin(course_value)].UniqueName.unique()
-            )
-        else:
+        if not course_value:
             return generate_empty_response(), display_style
+        exercise_value = select_all_exercises_in_course(local_df, course_value)
 
     # Filter dataframe
-    filtered_df = local_df
-    filtered_df = filtered_df[
-        (filtered_df.Attempt.ge(slider_value[0]))
-        & (filtered_df.Attempt.le(slider_value[1]))
-    ]
+    filtered_df = filter_attempt_limits(local_df, slider_value)
 
-    if course_value:
-        filtered_df = filtered_df[filtered_df.CourseName.isin(course_value)]
-
-    # We want all excercise values if Übersicht is checked
-    if "Übersicht" not in exercise_value:
-        if exercise_value:
-            filtered_df = filtered_df[filtered_df.UniqueName.isin(exercise_value)]
     if "Date" in checklist_value:
-        filtered_df = filtered_df[
-            (filtered_df.Time >= date_time_from) & (filtered_df.Time < date_time_to)
-        ]
+        filtered_df = filter_time(
+            filtered_df, convert_time(date_time_from), convert_time(date_time_to)
+        )
 
     if "Übersicht" in exercise_value:
-        hist_df = filtered_df[
-            [
-                "Joins",
-                "Projection_Attributes",
-                "Selection_Attributes",
-                "GroupBy",
-                "OrderBy",
-                "Strings",
-                "Tables",
-            ]
-        ]
+        hist_df = filtered_df[labels]
 
         if hist_df.empty:
             return generate_empty_response(), display_style
@@ -455,166 +442,30 @@ def update_histogram(
             }
 
         result_df = pd.DataFrame(result_dict)
-        # Colors for each bar
-        colors = [
-            "#60a7ba",
-            "#f0912d",
-            "#357025",
-            "#ba3622",
-            "#8f33d6",
-            "#6a4c4d",
-            "#cf8af3",
-        ]
 
-        # Label for each bar
-        labels = [
-            "Joins",
-            "Projection_Attributes",
-            "Selection_Attributes",
-            "GroupBy",
-            "OrderBy",
-            "Strings",
-            "Tables",
-        ]
-        values = []
+        local_df = get_values_from_data(result_df, labels)
 
-        for i, column in enumerate(result_df.columns):
-            values.append(result_df[column].Incorrect)
-
-        data = {"labels": labels, "values": values}
-        local_df = pd.DataFrame(data)
-
-        values = local_df["values"].astype(float)
-        local_df["values"] = (values * 100).astype(int)
-
-        fig = px.bar(
-            local_df,
-            x="labels",
-            y="values",
-            color="labels",
-            color_discrete_sequence=colors,
-            labels={
-                "labels": "SQL-Attribute",
-                "values": "STUDENTS",
-            },
-        )
+        fig = create_average_bar(local_df)
         fig.update_layout(showlegend=False, height=600)
         return fig, display_style
 
+    filtered_df = filtered_df[filtered_df.UniqueName.isin(exercise_value)]
     # Create figure which will be returned and contains all subfigures
-    fig = make_subplots(
-        rows=math.ceil(len(exercise_value) / 2),
-        cols=2,
-        subplot_titles=exercise_value,
-        shared_xaxes=True,
-        shared_yaxes=True,
-    )
+    fig = create_figure_with_subfigure(exercise_value)
 
     # Initialize subfigures
     for row in range(1, math.ceil(len(exercise_value) / 2) + 1):
         for col in range(1, 2 + 1):
             fig.append_trace(go.Bar(x=[], y=[]), row=row, col=col)
     # Order columns
-    hist_df = filtered_df[
-        [
-            "UniqueName",
-            "Joins",
-            "Projection_Attributes",
-            "Selection_Attributes",
-            "GroupBy",
-            "OrderBy",
-            "Strings",
-            "Tables",
-        ]
-    ].set_index("UniqueName")
+    hist_df = reduce_data_to_necessary_columns(filtered_df)
     # No data
     if hist_df.empty:
         return generate_empty_response(), display_style
 
-    # Colors for each bar
-    colors = [
-        "#60a7ba",
-        "#f0912d",
-        "#357025",
-        "#ba3622",
-        "#8f33d6",
-        "#6a4c4d",
-        "#cf8af3",
-    ]
-
-    # Label for each bar
-    labels = [
-        "Joins",
-        "Projection_Attributes",
-        "Selection_Attributes",
-        "GroupBy",
-        "OrderBy",
-        "Strings",
-        "Tables",
-    ]
-
-    # Each unique index represents one Course
-    for row, idx in enumerate(hist_df.index.unique()):
-        show_legend = row == 0
-        traces = []
-
-        # Calculate subfigure position in figure
-        row = (row + 1) / 2
-        col = 1 if row.is_integer() else 0
-        row = math.ceil(row) - 1
-
-        # Calculate dataframe for plot
-        task_subset_df = hist_df.loc[idx]
-        task_subset_df = task_subset_df.apply(pd.value_counts).T
-        task_subset_df = task_subset_df.div(task_subset_df.sum(axis=1), axis=0)
-
-        # Handle case if there are only correct answers
-        if task_subset_df.shape != (
-            7,
-            2,
-        ):  # sometimes task_subset_df is in the wrong shape
-            if task_subset_df.shape != (
-                7,
-                1,
-            ):
-                task_subset_df = task_subset_df.T
-
-            if "correct" in task_subset_df.columns.values:
-                task_subset_df["incorrect"] = 0
-
-        # Each bar needs a color and a legend entry and will therefore
-        # be plotted individually
-        for i, color in enumerate(colors):
-            trace = go.Bar(
-                x=[task_subset_df.index.values[i]],
-                y=[task_subset_df.incorrect[i] * 100],
-                name=labels[i],
-                marker={"color": color},
-                showlegend=show_legend,
-            )
-            traces.append(trace)
-
-            # All traces build one subfigure
-        for trace in traces:
-            fig.append_trace(trace, row=row + 1, col=col + 1)
-
-        # Figure styling
-        fig.update_layout(
-            height=400 * (row + 1),
-            legend={
-                "orientation": "h",
-                "xanchor": "left",
-                "yanchor": "bottom",
-                "x": 0.15,
-                "y": 1.05,
-            },
-        )
-        fig.update_xaxes(showticklabels=False)
-
-        for i in range(0, row + 1):
-            fig.update_yaxes(title_text="Students", row=i + 1, col=1)
-
+    fig = create_course_bars(hist_df, fig, labels)
     return fig, display_style
 
 
+# pylint: enable=too-many-locals
 # pylint: enable=too-many-arguments
