+++ modules/fbs-eat/eat/api/data_table/table.py
@@ -16,6 +16,8 @@ from api.util.utilities import (
     add_checklist,
     convert_time,
     create_invisible_time_row,
+    create_time_row,
+    filter_time,
 )
 
 
@@ -41,8 +43,9 @@ layout = html.Div(
             dbc.Card(
                 [
                     dbc.Col(
-                        timerow := html.Div(
-                            [
+                        html.Div(
+                            id="timerow_table",
+                            children=[
                                 "Date/Time From",
                                 dcc.Input(
                                     id="date_time_from_table",
@@ -57,7 +60,7 @@ layout = html.Div(
                                     ).strftime("%Y-%m-%dT%H:%M"),
                                     type="datetime-local",
                                 ),
-                            ]
+                            ],
                         ),
                     ),
                     add_checklist(
@@ -140,7 +143,7 @@ layout = html.Div(
 
 
 @callback(
-    Output(timerow, "children"),
+    Output("timerow_table", "children"),
     Output("is_date_on", "data"),
     Input("toggle_hiding_queries", "value"),
     Input("is_date_on", "data"),
@@ -152,110 +155,19 @@ def hide_date(date_hider, is_date_on):
     :param is_date_on: information if the date input was previously active
     :return: html Element for date input or have the input hidden
     """
-    if dash.callback_context.triggered[0]["value"] is None:
-        return (
-            html.Div(
-                [
-                    "Date/Time From",
-                    dcc.Input(
-                        id="date_time_from_table",
-                        value=datetime.now().strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                    "Date/Time To",
-                    dcc.Input(
-                        id="date_time_to_table",
-                        value=(
-                            datetime.now() + timedelta(hours=1, minutes=30)
-                        ).strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                ]
-            ),
-            True,
-        )
-    if (
-        "Exclude Date" not in dash.callback_context.triggered[0]["value"]
-        and is_date_on
-        and (len(date_hider) == 1 or not date_hider)
-    ):
-        return dash.no_update
-    if (
-        "Exclude equal queries" in dash.callback_context.triggered[0]["value"]
-        and not is_date_on
-        and (len(date_hider) == 1 or not date_hider)
-    ):
-        return (
-            html.Div(
-                [
-                    "Date/Time From",
-                    dcc.Input(
-                        id="date_time_from_table",
-                        value=datetime.now().strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                    "Date/Time To",
-                    dcc.Input(
-                        id="date_time_to_table",
-                        value=(
-                            datetime.now() + timedelta(hours=1, minutes=30)
-                        ).strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                ]
-            ),
-            True,
-        )
-    if not date_hider:
-        return (
-            html.Div(
-                [
-                    "Date/Time From",
-                    dcc.Input(
-                        id="date_time_from_table",
-                        value=datetime.now().strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                    "Date/Time To",
-                    dcc.Input(
-                        id="date_time_to_table",
-                        value=(
-                            datetime.now() + timedelta(hours=1, minutes=30)
-                        ).strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                ]
-            ),
-            True,
-        )
     if "Exclude Date" in date_hider:
         return (
             create_invisible_time_row("date_time_from_table", "date_time_to_table"),
             False,
         )
-    if "Exclude Date" not in date_hider:
-        return (
-            html.Div(
-                [
-                    "Date/Time From",
-                    dcc.Input(
-                        id="date_time_from_table",
-                        value=datetime.now().strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                    "Date/Time To",
-                    dcc.Input(
-                        id="date_time_to_table",
-                        value=(
-                            datetime.now() + timedelta(hours=1, minutes=30)
-                        ).strftime("%Y-%m-%dT%H:%M"),
-                        type="datetime-local",
-                    ),
-                ]
-            ),
-            True,
-        )
-    return dash.no_update
+
+    if "Exclude Date" not in date_hider and is_date_on:
+        return dash.no_update
+
+    return (
+        create_time_row("timerow_table", "date_time_from_table", "date_time_to_table"),
+        True,
+    )
 
 
 # Update date_time_to_table based on date_time_from_table
@@ -291,27 +203,26 @@ def read_query(query, date_time_from_table, date_time_to_table, daten, toggle_qu
         Printable String to show the user which filters are set. Also works for hided columns.
     """
 
-    date_time_from_table, date_time_to_table = convert_time(
-        date_time_from_table, date_time_to_table
-    )
-
     local_df = pd.read_json(daten)
     local_df["Time"] = pd.to_datetime(local_df["Time"])
 
     if toggle_queries:
+        if "Exclude Date" not in toggle_queries:
+            local_df = filter_time(
+                local_df,
+                convert_time(date_time_from_table),
+                convert_time(date_time_to_table),
+            )
         if "Exclude equal queries" in toggle_queries:
             # pylint: disable-next=no-member
             local_df = local_df.drop_duplicates(subset="Statement")
-        if "Exclude Date" not in toggle_queries:
-            local_df = local_df[
-                (local_df.Time >= date_time_from_table)
-                & (local_df.Time < date_time_to_table)
-            ]
     else:
-        local_df = local_df[
-            (local_df["Time"] >= date_time_from_table)
-            & (local_df["Time"] < date_time_to_table)
-        ]
+        local_df = filter_time(
+            local_df,
+            convert_time(date_time_from_table),
+            convert_time(date_time_to_table),
+        )
+
     tooltip_data = [
         {"Statement": str(row["Statement"])} for _, row in local_df.iterrows()
     ]
