+++ modules/fbs-eat/eat/api/util/utilities.py
@@ -6,10 +6,8 @@ from dash import html, dcc
 import dash_bootstrap_components as dbc
 import pandas as pd
 
+
 import dateutil.parser
-
-
-IMAGE_PATH = "assets/x.webp"
 
 
 def update_date_time(input_value, difference):
@@ -23,162 +21,6 @@ def update_date_time(input_value, difference):
     """
     date_time = dateutil.parser.parse(input_value)
     return date_time + timedelta(minutes=difference)
-
-
-def create_new_columns_buttons(columns):
-    """
-    create the buttons that display which columns are displayed
-    :return: a list of all buttons
-    """
-    columnbuttons = []
-    for _, column in enumerate(columns):
-        newcolbutton = html.Button(
-            id={"type": "delete-button-column", "index": column},
-            className="delete-button-column",
-            children=[
-                column,
-                html.Img(
-                    src=IMAGE_PATH, width="12", height="12", style={"margin": "7px"}
-                ),
-            ],
-        )
-        columnbuttons.append(newcolbutton)
-    return columnbuttons
-
-
-def create_new_filter_buttons(correctfilters, incorrectfilters):
-    """
-    create the buttons that display which filters are active
-    :return: a list of all buttons
-    """
-    correctfiltersbuttons = []
-    for _, correct_filter in enumerate(correctfilters):
-        newposbutton = html.Button(
-            id={"type": "delete-button-active", "index": correct_filter},
-            className="delete-button-active",
-            children=[
-                correct_filter,
-                html.Img(
-                    src=IMAGE_PATH, width="12", height="12", style={"margin": "7px"}
-                ),
-            ],
-        )
-        correctfiltersbuttons.append(newposbutton)
-    for _, incorrect_filter in enumerate(incorrectfilters):
-        newnegbutton = html.Button(
-            id={"type": "delete-button-active", "index": incorrect_filter},
-            className="delete-button-negative",
-            children=[
-                incorrect_filter,
-                html.Img(
-                    src=IMAGE_PATH, width="12", height="12", style={"margin": "7px"}
-                ),
-            ],
-        )
-        correctfiltersbuttons.append(newnegbutton)
-    return correctfiltersbuttons
-
-
-def prepare_data_for_graph(local_df, columns):
-    """
-    prepare the data to be used for making a graph
-    :param local_df: filtered data
-    :return: numbers for each attribute and their names
-    """
-    data = []
-    names = []
-    for i in range(0, 2 ** (len(columns))):
-        num = bin(i)
-        text = ""
-        requirements = []
-        length = len(columns)
-        for j in range(0, length):
-            binary_string = num[2:]
-            if len(binary_string) < length - j:
-                requirements.append("incorrect")
-                text = text + " " + columns[j] + " " + "incorrect"
-            else:
-                if binary_string[j - length] == "1":
-                    requirements.append("correct")
-                    text = text + " " + columns[j] + " " + "correct"
-                else:
-                    requirements.append("incorrect")
-                    text = text + " " + columns[j] + " " + "incorrect"
-        names.append(text)
-        tmplocal_df = local_df
-        for k, single_column in enumerate(columns):
-            tmplocal_df = tmplocal_df[
-                tmplocal_df[single_column.replace(" ", "_")] == requirements[(int)(k)]
-            ]
-        data.append(len(tmplocal_df.index))
-    return data, names
-
-
-def create_marks_for_slider(local_df, all_filters):
-    """
-    calculate how many attempts have all filters correct and incorrect
-    :param local_df: the filtered data
-    :return: the marks to add on the sliders
-    """
-    sliders = []
-    for row in all_filters:
-        tmprow = row.replace(" ", "_")
-        tmp_local_df = local_df
-        tmpdf_correct = tmp_local_df[tmp_local_df[tmprow] == "correct"]
-        tmp_local_df = local_df
-        tmpdf_incorrect = tmp_local_df[tmp_local_df[tmprow] == "incorrect"]
-        marks = {
-            0: str(len(tmpdf_correct.index)),
-            1: "",
-            2: str(len(tmpdf_incorrect.index)),
-        }
-        sliders.append(marks)
-    return sliders
-
-
-def update_button(button, trigger, all_filters):
-    """
-    update the buttons after they were selected or a slider was changed
-    :param button: current buttons
-    :param trigger: trigger for the callback
-    :return: new buttons
-    """
-    if "add-button" in trigger:
-        for counter, all_buttons in enumerate(all_filters):
-            if all_buttons in trigger:
-                if button[counter] == "+":
-                    button[counter] = "-"
-                else:
-                    button[counter] = "+"
-
-    if "delete-button-column" in trigger:
-        for counter, active_filter in enumerate(all_filters):
-            if active_filter in trigger:
-                button[counter] = "+"
-    if "slider" in trigger:
-        for count, filtername in enumerate(all_filters):
-            if filtername in trigger:
-                button[count] = "+"
-    return button
-
-
-def update_slider(slider, trigger, all_filters):
-    """
-    update the sliders if an attribute was selected
-    :param slider: current sliders
-    :param trigger: trigger for the callback
-    :return: new sliders
-    """
-    if "add-button" in trigger:
-        for count, filtername in enumerate(all_filters):
-            if filtername in trigger:
-                slider[count] = 1
-
-    if "delete-button-active" in trigger:
-        for counter, active_columns in enumerate(all_filters):
-            if active_columns in trigger:
-                slider[counter] = 1
-    return slider
 
 
 def filter_checklist(checks, local_df, slider_id):
@@ -433,13 +275,53 @@ def filter_data(data):
     return local_df
 
 
-def convert_time(date_time_from, date_time_to):
+def filter_attempt_limits(local_df, limits):
+    """
+    filter data to all entries within the attempt limit
+    :param data: all data
+    :param limits: the minimum and maximum amout of attempts
+    :return: data without entreis that have too many or to few attempt numbers
+    """
+    limits = limits or [local_df.Attempt.min(), local_df.Attempt.max()]
+    local_df = local_df[
+        (local_df.Attempt.ge(limits[0])) & (local_df.Attempt.le(limits[1]))
+    ]
+    return local_df
+
+
+def select_all_exercises_in_course(local_df, course_value):
+    """
+    select all exercises from a course if no exercises are selected
+    :param local_df: current filtered data
+    :param course_value: selected courses
+    :return: all exercises in the selected courses as a list
+    """
+    exercise_value = list(
+        local_df[local_df.CourseName.isin(course_value)].UniqueName.unique()
+    )
+    return exercise_value
+
+
+def convert_time(date_time):
     """
     converts the format of two given times
     :param date_time_from: time to be converted
     :param date_time_to: time to be converted
     :return: times with converted format
     """
-    date_time_from = dateutil.parser.parse(date_time_from)
-    date_time_to = dateutil.parser.parse(date_time_to)
-    return date_time_from, date_time_to
+    date_time = dateutil.parser.parse(date_time)
+    return date_time
+
+
+def filter_time(data_frame, date_time_from, date_time_to):
+    """
+    filter the data for entries in a given timeframe
+    :param data_frame: data not yet filtered for time
+    :param date_time_from: start date for the timeframe
+    :param date_time_to: end date for the timeframe
+    :return: data filtered for time frame
+    """
+    data_frame = data_frame[
+        (data_frame.Time >= date_time_from) & (data_frame.Time < date_time_to)
+    ]
+    return data_frame
